some refactoring, rename extension name and package to Pixiv Comic
