+++ src/ja/pixivcomic/res/mipmap-mdpi/ic_launcher.png
