+++ src/ja/pixivcomic/res/mipmap-xxhdpi/ic_launcher.png
